requirements.txt
fastapi
uvicorn[standard]
sqlite-utils
pydantic
python-dotenv
pytest
openai>=1.0
psycopg2-binary
sqlalchemy
alembic